# nix/outputs/templates.spring-rest-api/src/main/java/com/example/demo/endpoints/api/TasksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.demo.endpoints.api;

import com.example.demo.endpoints.model.NewTask;
import com.example.demo.endpoints.model.Task;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-08T08:54:10.338848230-06:00[America/Denver]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "tasks", description = "the tasks API")
public interface TasksApi {

    String PATH_CREATE_TASK = "/tasks";
    /**
     * POST /tasks : Create a new task
     * Adds a new task to the manager.
     *
     * @param newTask  (required)
     * @return Task created successfully. (status code 201)
     *         or Invalid input provided. (status code 400)
     */
    @Operation(
        operationId = "createTask",
        summary = "Create a new task",
        description = "Adds a new task to the manager.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Task created successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input provided.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TasksApi.PATH_CREATE_TASK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<Task> createTask(
        @Parameter(name = "NewTask", description = "", required = true) @Valid @RequestBody NewTask newTask
    );


    String PATH_GET_TASK_BY_ID = "/tasks/{taskId}";
    /**
     * GET /tasks/{taskId} : Get task details
     * Retrieves a single task by its unique ID.
     *
     * @param taskId The ID of the task to retrieve. (required)
     * @return Task found. (status code 200)
     *         or Task not found. (status code 404)
     */
    @Operation(
        operationId = "getTaskById",
        summary = "Get task details",
        description = "Retrieves a single task by its unique ID.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Task found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            }),
            @ApiResponse(responseCode = "404", description = "Task not found.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TasksApi.PATH_GET_TASK_BY_ID,
        produces = { "application/json" }
    )
    ResponseEntity<Task> getTaskById(
        @Parameter(name = "taskId", description = "The ID of the task to retrieve.", required = true, in = ParameterIn.PATH) @PathVariable("taskId") UUID taskId
    );


    String PATH_GET_TASKS = "/tasks";
    /**
     * GET /tasks : Retrieve a list of tasks
     * Returns a JSON array containing all tasks.
     *
     * @return A successful response. (status code 200)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "getTasks",
        summary = "Retrieve a list of tasks",
        description = "Returns a JSON array containing all tasks.",
        responses = {
            @ApiResponse(responseCode = "200", description = "A successful response.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Task.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TasksApi.PATH_GET_TASKS,
        produces = { "application/json" }
    )
    ResponseEntity<List<Task>> getTasks(

    );

}