# nix/outputs/templates.spring-rest-api/src/main/openapi/openapi-generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Manager API",
    "description": "A sample API for managing user tasks and to-do lists.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Retrieve a list of tasks",
        "operationId": "getTasks",
        "description": "Returns a JSON array containing all tasks.",
        "responses": {
          "200": {
            "description": "A successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "operationId": "createTask",
        "description": "Adds a new task to the manager.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTask"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input provided."
          }
        }
      }
    },
    "/tasks/{taskId}": {
      "get": {
        "summary": "Get task details",
        "operationId": "getTaskById",
        "description": "Retrieves a single task by its unique ID.",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "description": "The ID of the task to retrieve.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "required": ["id", "title", "completed"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "d3b07384-d113-4956-a5cc-484194fc842d"
          },
          "title": {
            "type": "string",
            "example": "Buy groceries"
          },
          "description": {
            "type": "string",
            "example": "Need to pick up milk, eggs, and bread."
          },
          "completed": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-31T20:18:00Z"
          }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Buy groceries"
          },
          "description": {
            "type": "string",
            "example": "Need to pick up milk, eggs, and bread."
          },
          "completed": {
            "type": "boolean",
            "default": false
          }
        }
      }
    }
  }
}